\name{deer}
\alias{deer}
\docType{data}
\title{Giant deer premolars}
\description{
  Breadth of the lower left second premolar of the giant deer
  (\emph{Megaloceros pachyoseus})
}
\usage{data(deer)}
\format{
  'deer' is a vector with 39 observations on the lower left
  second premolar of the giant deer
}
\details{
 
}

\source{
Dong, Z. (1996) Looking into Peking Man's subsistence - a taphonomic
analysis of the middle pleistocene \emph{Homo erectus} site in China.
Unpublished PhD thesis, Department of Anthropology, Indiana University   
}
\references{
  Flury, B.D. (1997) \emph{A First Course in Multivariate Statistics},
  New York: Springer
}

\examples{
data(deer)
\dontrun{
hist(deer, breaks = 10, main = "Deer lower left 2nd premolar")}
}
\keyword{datasets}
