\name{dogwood}
\alias{dogwood}
\docType{data}
\title{Flowering dogwood trees}
\description{
  Data collected on flowering dogwood trees collected in 39 equally
  sized plots in Hoosier National Forest in Southern Indiana.   The
  total number of trees, the number flowering and the aspect (direction
  of slope) of the plot were recorded.
}
\usage{data(dogwood)}
\format{
  'dogwood' is a data frame with observations from 39 plots on the following 3 variables.
  \describe{
    \item{\code{Aspect}}{Orientation of the plot, i.e. direction of
      steepest descent, in degrees clockwise from North}
    \item{\code{Total.Trees}}{Total number of trees in the plot}
    \item{\code{Flowering.Trees}}{Number of flowering dogwood trees in
      the plot}
  }
}
\details{

}
\source{
Data courtesy of K. Dorman and D.Whitehead, Indiana University
Department of Biology
}
\references{
   Flury, B.D. (1997) \emph{A First Course in Multivariate Statistics},
New York: Springer
}
\examples{
data(dogwood)

}
\keyword{datasets}
