\name{Flury-package}
\alias{Flury-package}
\alias{Flury}
\docType{package}
\title{
Data sets from Flury (1997)
}
\description{
This package contains all the datasets included in Flury (1997), and is
intended to support use of the book when teaching with \R.
}
\details{
\tabular{ll}{
Package: \tab Flury\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2006-06-15\cr
License: \tab As per the datasets supplied on the publishers website\cr
}

}
\author{
Bernard Flury

Maintainer: Paul Hewson <paul.hewson@plymouth.ac.uk>

}
\references{Flury (1997) \emph{A First Course in Multivariate
  Statistics} New York: Springer  ISBN 0-387-98206-X
}
\keyword{ package }

\seealso{\code{lookup} cross references the tables in the book with the \R
  object names

}
\examples{

}
