\name{lookup}
\alias{lookup}
\docType{data}
\title{Lookup table}
\description{
  This cross references the data objects and the tables in Flury (1997)
}
\usage{data(lookup)}
\format{
  A data frame with 33 observations on the following variable.
  \describe{
    \item{\code{Data.Name}}{R data object name}
  }
}
\details{

}
\source{

}
\references{
   Flury, B.D. (1997) \emph{A First Course in Multivariate Statistics},
New York: Springer
}
\examples{
data(lookup)
\dontrun{lookup}
}
\keyword{datasets}
